copying validation moved to result view
